Finish impl of dinosaur name generation using RNN
It's finally working! I can't believe I missed passing in the
previous activations to the RNN at each timestep. This was
effectively erasing all the learning because the activations are
be reset to zero by default when not passed! Such a dumb mistake!
But a great learning on how to implement RNN's with PyTorch :)